VOICE clone/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
PyPDF2==3.0.1
numpy==1.24.3
torch==2.1.0
torchaudio==2.1.0
librosa==0.10.1
soundfile==0.12.1
python-jose==3.3.0
pydantic==2.4.2
aiohttp==3.9.1
groq==0.4.1
gTTS==2.4.0 